Made image quality better
Changed takeImage.js to start the video stream when the task is opened and keep it running throughout the whole task. Previously, the stream was starting and stopping each time take_image() was called which didn't give the camera enouch time to adapt to the correct exposure.
I also changed takeImage.js to have smaller functions that make more sense
In mouseAndTouchEvent.js, I got rid of a delay in fixation durations and deleted unecissary commented code.
In Index.html, I made it so that the screen that is up when the image is being taken is black on top and white on bottom to reduce glare.